services/photo-service/requirements.txt
fastapi
uvicorn
pydantic<2.0.0
python-multipart
pymysql
sqlalchemy
alembic
Pillow
python-dotenv
boto3==1.35.54
botocore==1.35.54
requests
pytest
httpx
